[TD-19] Global instance settings with automatic hot reload (#19)
Common utility to manage settings shared by all componets in instance.

Settings are seeked following this strategy:

- TD_\<KEY> environment variable
- TD__\<INSTANCE>_\<KEY> environment variable
- Entry \<key> in file settings.yaml in instance folder.
- Entry \<key> in default bundled file settings.yaml.

If a file named settings_\<instance>.yaml is found in .tabsdata folder, it will be moved to the instance folder when first loading settings, renaming it to expected file settings.yaml.

This component supports hot reload for all the possible sources of values as described above.