[TD-648] Inter-collection read permission is not propagated as read permission to the role (#563)
If a role has admin/dev permissions on a collection X, the any collection that has the the 
collection X with inter-collection permission must be considered as having read permission 
for the role.


Adding the authz check for InterCollection reads to donwload, list, list_data_versions, 
sample and schema services.